projects/analytics-engineering/job-market-intelligence/pipeline/requirements.txt
# Job Market Intelligence - Python Dependencies

# Web Scraping
scrapy==2.11.0
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# NLP & Text Processing
spacy==3.7.2
# Run: python -m spacy download en_core_web_sm

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Scheduling & Task Queue
apache-airflow==2.7.3
celery==5.3.4
redis==5.0.1

# Dashboard
streamlit==1.28.2
plotly==5.18.0
altair==5.1.2

# Notifications
python-telegram-bot==20.7
sendgrid==6.11.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0


